requirements.txt
python-miio==0.4.7